+++ src/builtins/chainables/attribute.ts
@@ -8,8 +8,8 @@ import {
   Product,
   ProductType,
   ProductTypeish,
-} from "../core/mod.ts";
-import { ValueWrapperProduct } from "./ValueWrapperProduct.ts";
+} from "../../core/mod.ts";
+import { ValueWrapperProduct } from "../ValueWrapperProduct.ts";
 
 type StringKeys<T> = `${Exclude<keyof T, symbol>}`;
 type GetStringKey<T, K extends StringKeys<T>> = K extends keyof T ? T[K]
