+++ src/builtins/chainables/convert.ts
@@ -8,7 +8,7 @@ import {
   Product,
   ProductType,
   ProductTypeish,
-} from "../core/mod.ts";
+} from "../../core/mod.ts";
 
 export interface ConvertArgsHkt extends Hkt<[Product]> {
   [Hkt.output]: [productTypeish: ProductTypeish<Hkt.Input<this>[0]>];
