+++ src/builtins/chainables/mod.ts
@@ -87,3 +87,27 @@ export default {
   sphere,
   polyhedron,
 };
+
+// moderate
+
+export * from "./attribute.ts";
+export * from "./convert.ts";
+export * from "./cube.ts";
+export * from "./cylinder.ts";
+export * from "./diff.ts";
+export * from "./flip.ts";
+export * from "./getBoundingBox.ts";
+export * from "./intersection.ts";
+export * from "./label.ts";
+export * from "./meld.ts";
+export * from "./mirror.ts";
+export * from "./moveTo.ts";
+export * from "./polyhedron.ts";
+export * from "./reflect.ts";
+export * from "./rotate.ts";
+export * from "./scale.ts";
+export * from "./shift.ts";
+export * from "./sphere.ts";
+export * from "./translate.ts";
+export * from "./union.ts";
+export * from "./unionDifference.ts";
