+++ src/builtins/chainables/shift.ts
@@ -1,6 +1,6 @@
-import { Component, ConvertibleElement, Operation } from "../core/mod.ts";
-import { interpretTriplet, Triplet } from "./helpers.ts";
-import { Mesh } from "./Mesh.ts";
+import { Component, ConvertibleElement, Operation } from "../../core/mod.ts";
+import { interpretTriplet, Triplet } from "../helpers.ts";
+import { Mesh } from "../Mesh.ts";
 import { moveTo } from "./moveTo.ts";
 
 export const shift = Component.create("shift", (triplet: Triplet) => {
