+++ src/builtins/chainables/sphere.ts
@@ -1,4 +1,4 @@
-import { Mesh } from "./Mesh.ts";
+import { Mesh } from "../Mesh.ts";
 import {
   Component,
   Conversion,
@@ -7,11 +7,11 @@ import {
   Element,
   Id,
   LeafProduct,
-} from "../core/mod.ts";
-import { interpretTriplet, Triplet } from "./helpers.ts";
-import { Smooth, smoothContext } from "./smoothContext.ts";
-import { Face } from "./Face.ts";
-import { Vector3 } from "./Vector3.ts";
+} from "../../core/mod.ts";
+import { interpretTriplet, Triplet } from "../helpers.ts";
+import { Smooth, smoothContext } from "../smoothContext.ts";
+import { Face } from "../Face.ts";
+import { Vector3 } from "../Vector3.ts";
 
 const tau = Math.PI * 2;
