+++ src/builtins/mod.ts
@@ -4,10 +4,11 @@ registerPlugin(
   new URL("../client-builtins/mod.ts", import.meta.url).toString(),
 );
 
-export { default } from "./chainables.ts";
+export { default } from "./chainables/mod.ts";
 
 // moderate --exclude register.ts --exclude helpers.ts
 
+export * from "./chainables/mod.ts";
 export * from "./serializers/mod.ts";
 export * from "./BooleanParam.ts";
 export * from "./BoundingBox.ts";
@@ -20,28 +21,6 @@ export * from "./Plane.ts";
 export * from "./Transformation.ts";
 export * from "./ValueWrapperProduct.ts";
 export * from "./Vector3.ts";
-export * from "./attribute.ts";
 export * from "./bspMeshConversion.ts";
-export * from "./chainables.ts";
-export * from "./convert.ts";
-export * from "./cube.ts";
-export * from "./cylinder.ts";
-export * from "./diff.ts";
-export * from "./flip.ts";
-export * from "./getBoundingBox.ts";
-export * from "./intersection.ts";
-export * from "./label.ts";
-export * from "./meld.ts";
-export * from "./mirror.ts";
-export * from "./moveTo.ts";
-export * from "./polyhedron.ts";
-export * from "./reflect.ts";
-export * from "./rotate.ts";
-export * from "./scale.ts";
-export * from "./shift.ts";
 export * from "./smoothContext.ts";
-export * from "./sphere.ts";
 export * from "./stl.ts";
-export * from "./translate.ts";
-export * from "./union.ts";
-export * from "./unionDifference.ts";
